Cleaned up the code a LOT.
Made a new function to check if a file exists, as I had this copied-and-pasted code in a few other functions.

Turned the "morning_questions()" and "evening_questions()" functions into 1 function.

Replaced the horrible "if-elif" loop of "day_recall()" with a much better for loop.

Turned the mornings_prompts into 1 list of prompts, rather than a variable for each statement. Did the same with evening_prompts too.